Moved urls to appropriate apps
