+++ misc/urls/practical-urls.py
@@ -1,0 +1,6 @@
+from django.conf.urls import url, include
+from misc import views
+
+urlpatterns = [
+    url(r'^$', views.Practical.as_view(), name='practical'),
+]
